And move ClientException into new exceptions module too
